Linux compile problem elimination
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1486 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/nolong/rexxapi/unix/SUBCOMCommand.cpp
@@ -78,7 +78,7 @@
 #define SECOND_PARAMETER 0             /* 0 for no  NL_CAT_LOCALE    */
 #endif
 
-void parmerr(ULONG);
+void parmerr(size_t); 
 
 int main(                              /* Program Begins             */
 int     argc,                          /* Argument count             */
